refactor/Remove warnings at BahrainOBF
