+++ obp-api/src/main/scala/code/api/BahrainOBF/v1_0_0/BeneficiariesApi.scala
@@ -1,27 +1,19 @@
 package code.api.BahrainOBF.v1_0_0
 
-import code.api.APIFailureNewStyle
 import code.api.berlin.group.v1_3.JvalueCaseClass
-import net.liftweb.json
-import net.liftweb.json._
-import code.api.util.APIUtil.{defaultBankId, _}
-import code.api.util.NewStyle
-import code.api.util.ErrorMessages._
+import code.api.util.APIUtil._
+import code.api.util.ApiTag
 import code.api.util.ApiTag._
-import code.api.util.NewStyle.HttpCode
-import code.bankconnectors.Connector
-import code.model._
-import code.util.Helper
-import code.views.Views
+import code.api.util.ErrorMessages._
+import com.github.dwickern.macros.NameOf.nameOf
+import com.openbankproject.commons.ExecutionContext.Implicits.global
 import net.liftweb.common.Full
 import net.liftweb.http.rest.RestHelper
-import com.github.dwickern.macros.NameOf.nameOf
+import net.liftweb.json
+import net.liftweb.json._
+
 import scala.collection.immutable.Nil
 import scala.collection.mutable.ArrayBuffer
-import com.openbankproject.commons.ExecutionContext.Implicits.global
-import scala.concurrent.Future
-import code.api.BahrainOBF.v1_0_0.ApiCollector
-import code.api.util.ApiTag
 
 object APIMethods_BeneficiariesApi extends RestHelper {
     val apiVersion =  ApiCollector.apiVersion
